chore(repo): phase 1 cleanup and consistency fixes\n\n- Remove root example run tasks; keep in :streamconverter-examples\n- Fix runDemo main class and consolidate example run tasks\n- Prune root dependencies; align comments (Netty/Reactor)\n- Unify Javadoc output paths under build/docs/javadoc\n- Web stack: drop starter-web, standardize on WebFlux\n- Core: remove Boot-specific resources (application*.yml, logback*.xml)\n- Rename package dir charaCode -> charcode and update imports\n- Fix README example links to actual module paths\n- Add AGENTS.md contributor guide\n- Add scripts/create-issues.sh to open tracking issues via gh CLI\n\nRefs: #289, #290, #291, #292, #293, #294, #296, #297, #298, #299
