+++ streamconverter-core/build.gradle.kts
@@ -156,7 +156,7 @@ tasks.pitest {
 tasks.javadoc {
     options.encoding = "UTF-8"
     options.memberLevel = org.gradle.external.javadoc.JavadocMemberLevel.PROTECTED
-    setDestinationDir(file("docs/javadoc"))
+    setDestinationDir(file("$buildDir/docs/javadoc"))
 }
 
 // spotlessCheck タスクを無効化
