--- streamconverter-core/src/main/resources/application-prod.properties
@@ -1,89 +1,0 @@
-# StreamConverter Production Environment Security Configuration
-# 本番環境向けセキュリティ設定
-
-# ====================
-# セキュリティ強化設定
-# ====================
-
-# XML Security Settings - XML外部エンティティ攻撃防止
-security.xml.disable-external-entities=true
-security.xml.disable-doctype-declarations=true
-security.xml.load-external-dtd=false
-
-# XPath Security Settings - XPathインジェクション防止
-security.xpath.validation.enabled=true
-security.xpath.strict-mode=true
-
-# Path Traversal Prevention - パストラバーサル攻撃防止
-security.path-traversal.prevention=true
-security.path-traversal.allow-parent-references=false
-security.file-access.restrict-to-workspace=true
-
-# SQL Injection Prevention - SQLインジェクション防止
-security.sql.use-prepared-statements=true
-security.sql.validate-queries=true
-
-# Input Validation - 入力値検証強化
-security.input.max-file-size=100MB
-security.input.allowed-file-extensions=.json,.xml,.csv,.txt
-security.input.validate-encoding=true
-
-# ====================
-# ログ設定
-# ====================
-
-# Security Event Logging - セキュリティイベントのログ記録
-logging.security.enabled=true
-logging.security.level=WARN
-logging.security.file=logs/security.log
-
-# Sensitive Data Protection - 機密データ保護
-logging.mask-sensitive-data=true
-logging.exclude-request-body=true
-logging.exclude-response-body=true
-
-# ====================
-# パフォーマンス設定
-# ====================
-
-# Memory Management - メモリ管理
-performance.memory.max-heap-usage=80
-performance.memory.gc-after-large-operations=true
-
-# Stream Processing - ストリーミング処理
-performance.streaming.buffer-size=64KB
-performance.streaming.enable-compression=false
-
-# Timeout Settings - タイムアウト設定
-performance.timeout.command-execution=300000
-performance.timeout.file-processing=600000
-
-# ====================
-# 監査・モニタリング
-# ====================
-
-# Audit Logging - 監査ログ
-audit.enabled=true
-audit.log-file=logs/audit.log
-audit.include-user-actions=true
-audit.include-system-events=true
-
-# Performance Monitoring - パフォーマンス監視
-monitoring.enabled=true
-monitoring.metrics.memory=true
-monitoring.metrics.processing-time=true
-monitoring.metrics.throughput=true
-
-# ====================
-# 環境固有設定
-# ====================
-
-# Production Environment Flags
-environment.name=production
-environment.debug.enabled=false
-environment.test-mode.enabled=false
-
-# Security Headers
-security.headers.content-type-options=nosniff
-security.headers.frame-options=DENY
-security.headers.xss-protection=1; mode=block
