--- streamconverter-core/src/main/resources/application.yml
@@ -1,14 +1,0 @@
-server:
-  port: 8080
-  
-spring:
-  application:
-    name: stream-converter-api
-  main:
-    allow-bean-definition-overriding: true
-    
-logging:
-  level:
-    com.streamConverter: INFO
-    org.springframework.web: INFO
-    root: WARN
