--- streamconverter-core/src/main/resources/logback-spring.xml
@@ -1,133 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <!-- Profile-specific configurations -->
-    
-    <!-- Development Profile -->
-    <springProfile name="!prod">
-        <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-            <encoder>
-                <pattern>%d{yyyy-MM-dd HH:mm:ss} %-5level %logger{36} [%X{requestId:-}] [%X{extractedValue:-}] [%X{command:-}] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-            <file>logs/streamconverter.log</file>
-            <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-                <fileNamePattern>logs/streamconverter.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-                <maxHistory>7</maxHistory>
-                <maxFileSize>10MB</maxFileSize>
-                <totalSizeCap>100MB</totalSizeCap>
-            </rollingPolicy>
-            <encoder>
-                <pattern>%date %-5level %logger{36} [%X{requestId:-}] [%X{extractedValue:-}] [%X{command:-}] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <root level="INFO">
-            <appender-ref ref="STDOUT" />
-            <appender-ref ref="FILE" />
-        </root>
-    </springProfile>
-    
-    <!-- Production Profile -->
-    <springProfile name="prod">
-        <!-- Main Application Log -->
-        <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-            <file>logs/streamconverter.log</file>
-            <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-                <fileNamePattern>logs/streamconverter.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-                <maxHistory>30</maxHistory>
-                <maxFileSize>50MB</maxFileSize>
-                <totalSizeCap>2GB</totalSizeCap>
-            </rollingPolicy>
-            <encoder>
-                <pattern>%date %-5level %logger{36} [%X{requestId:-}] [%X{extractedValue:-}] [%X{command:-}] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <!-- Security Events Log -->
-        <appender name="SECURITY" class="ch.qos.logback.core.rolling.RollingFileAppender">
-            <file>logs/security.log</file>
-            <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-                <fileNamePattern>logs/security.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-                <maxHistory>90</maxHistory>
-                <maxFileSize>20MB</maxFileSize>
-                <totalSizeCap>500MB</totalSizeCap>
-            </rollingPolicy>
-            <encoder>
-                <pattern>%date %-5level %logger{36} [%X{requestId:-}] [SECURITY] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <!-- Audit Events Log -->
-        <appender name="AUDIT" class="ch.qos.logback.core.rolling.RollingFileAppender">
-            <file>logs/audit.log</file>
-            <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-                <fileNamePattern>logs/audit.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-                <maxHistory>365</maxHistory>
-                <maxFileSize>20MB</maxFileSize>
-                <totalSizeCap>1GB</totalSizeCap>
-            </rollingPolicy>
-            <encoder>
-                <pattern>%date %-5level %logger{36} [%X{requestId:-}] [AUDIT] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <!-- Performance Monitoring Log -->
-        <appender name="PERFORMANCE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-            <file>logs/performance.log</file>
-            <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-                <fileNamePattern>logs/performance.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-                <maxHistory>30</maxHistory>
-                <maxFileSize>20MB</maxFileSize>
-                <totalSizeCap>500MB</totalSizeCap>
-            </rollingPolicy>
-            <encoder>
-                <pattern>%date %-5level %logger{36} [%X{requestId:-}] [PERFORMANCE] - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <!-- Error-only Console Output for Production -->
-        <appender name="CONSOLE_ERROR" class="ch.qos.logback.core.ConsoleAppender">
-            <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
-                <level>ERROR</level>
-            </filter>
-            <encoder>
-                <pattern>%d{yyyy-MM-dd HH:mm:ss} %-5level %logger{36} - %msg%n</pattern>
-            </encoder>
-        </appender>
-        
-        <!-- Logger configurations for Production -->
-        <logger name="com.streamConverter.security" level="INFO" additivity="false">
-            <appender-ref ref="SECURITY" />
-            <appender-ref ref="FILE" />
-        </logger>
-        
-        <logger name="com.streamConverter.audit" level="INFO" additivity="false">
-            <appender-ref ref="AUDIT" />
-            <appender-ref ref="FILE" />
-        </logger>
-        
-        <logger name="com.streamConverter.benchmark" level="INFO" additivity="false">
-            <appender-ref ref="PERFORMANCE" />
-            <appender-ref ref="FILE" />
-        </logger>
-        
-        <logger name="com.streamConverter.monitor" level="INFO" additivity="false">
-            <appender-ref ref="PERFORMANCE" />
-            <appender-ref ref="FILE" />
-        </logger>
-        
-        <!-- Root logger for Production -->
-        <root level="WARN">
-            <appender-ref ref="FILE" />
-            <appender-ref ref="CONSOLE_ERROR" />
-        </root>
-    </springProfile>
-    
-    <!-- Fallback configuration (when no profile is active) -->
-    <springProfile name="!prod,!dev">
-        <include resource="logback.xml"/>
-    </springProfile>
-    
-</configuration>
