--- streamconverter-core/src/main/resources/logback.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<configuration>
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{yyyy-MM-dd HH:mm:ss} %-5level %logger{36} [%X{requestId:-}] [%X{extractedValue:-}] [%X{command:-}] - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <!-- 現在のログファイル -->
-        <file>logs/streamconverter.log</file>
-        <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-            <!-- 日付を含むファイル名パターン。例えば、streamconverter.2025-04-04.1.log -->
-            <fileNamePattern>logs/streamconverter.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-            <!-- 例えば過去30日分を保持し、それ以外は自動削除 -->
-            <maxHistory>30</maxHistory>
-            <!-- 1ファイルの最大サイズ -->
-            <maxFileSize>10MB</maxFileSize>
-            <!-- 全ログファイルの最大サイズ -->
-            <totalSizeCap>1GB</totalSizeCap>
-        </rollingPolicy>
-        <encoder>
-            <pattern>%date %-5level %logger{36} [%X{requestId:-}] [%X{extractedValue:-}] [%X{command:-}] - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-
-    <root level="INFO">
-        <appender-ref ref="STDOUT" />
-        <appender-ref ref="FILE" />
-    </root>
-
-</configuration>
-<!-- This is a simple Logback configuration file that outputs log messages to the console. 
-     The log format includes the date, log level, logger name, and the message itself. 
-     The root logger is set to INFO level, meaning it will log messages at INFO level and above. -->
