+++ streamconverter-tools/src/test/java/com/streamConverter/benchmark/MemoryEfficiencyQuickTest.java
@@ -2,7 +2,7 @@ package com.streamConverter.benchmark;
 
 import com.streamConverter.*;
 import com.streamConverter.command.IStreamCommand;
-import com.streamConverter.command.impl.charaCode.CharacterConvertCommand;
+import com.streamConverter.command.impl.charcode.CharacterConvertCommand;
 import java.io.*;
 import java.util.*;
 import org.junit.jupiter.api.*;
