+++ streamconverter-web/build.gradle.kts
@@ -19,7 +19,6 @@ dependencies {
     implementation(project(":streamconverter-core"))
     
     // Spring Boot Web dependencies
-    implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-webflux")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
